modify join and delete
