+++ cmd/sealer/cmd/cluster/join.go
@@ -17,11 +17,11 @@ package cluster
 import (
 	"fmt"
 	"io/ioutil"
-	"net"
 	"path/filepath"
 	"strings"
 
 	"github.com/sealerio/sealer/cmd/sealer/cmd/utils"
+	utilsnet "github.com/sealerio/sealer/utils/net"
 
 	"github.com/sealerio/sealer/apply"
 	"github.com/sealerio/sealer/common"
@@ -31,13 +31,13 @@ import (
 	"github.com/sealerio/sealer/pkg/imagedistributor"
 	"github.com/sealerio/sealer/pkg/imageengine"
 	"github.com/sealerio/sealer/pkg/infradriver"
-	utilsnet "github.com/sealerio/sealer/utils/net"
 	"github.com/spf13/cobra"
 )
 
 var clusterName string
 var joinArgs *apply.Args
-var newMasters, newWorkers []net.IP
+var newMasters string
+var newWorkers string
 
 var exampleForJoinCmd = `
 join default cluster:
@@ -58,6 +58,9 @@ func NewJoinCmd() *cobra.Command {
 				cf  clusterfile.Interface
 				err error
 			)
+			if err := utils.ValidateJoinArgs(newMasters, newWorkers); err != nil {
+				return fmt.Errorf("failed to validate input run args: %v", err)
+			}
 			workClusterfile := common.GetClusterWorkClusterfile()
 			clusterFileData, err := ioutil.ReadFile(filepath.Clean(workClusterfile))
 			if err != nil {
@@ -68,16 +71,8 @@ func NewJoinCmd() *cobra.Command {
 				return err
 			}
 			cluster := cf.GetCluster()
-			masterList := utilsnet.IPsToIPStrs(newMasters)
-			workerList := utilsnet.IPsToIPStrs(newWorkers)
-			joinMasters := strings.Join(masterList, ",")
-			joinNodes := strings.Join(workerList, ",")
 
-			if err := utils.ValidateJoinArgs(joinMasters, joinNodes); err != nil {
-				return fmt.Errorf("failed to validate input run args: %v", err)
-			}
-
-			if err := utils.Join(&cluster, joinArgs, joinMasters, joinNodes); err != nil {
+			if err := utils.JoinClusterNode(&cluster, joinArgs, newMasters, newWorkers); err != nil {
 				return err
 			}
 
@@ -98,9 +93,14 @@ func NewJoinCmd() *cobra.Command {
 				return err
 			}
 
+			masters := strings.Split(newMasters, ",")
+			masterIpList := utilsnet.IPStrsToIPs(masters)
+			nodes := strings.Split(newMasters, ",")
+			nodeIpList := utilsnet.IPStrsToIPs(nodes)
+
 			var (
 				clusterImageName = cluster.Spec.Image
-				hosts            = append(newMasters, newWorkers...)
+				hosts            = append(masterIpList, nodeIpList...)
 			)
 
 			// distribute rootfs
@@ -119,7 +119,7 @@ func NewJoinCmd() *cobra.Command {
 
 			installer, err := clusterruntime.NewInstaller(infraDriver, imageEngine, *runtimeConfig)
 
-			_, _, err = installer.ScaleUp(newMasters, newWorkers)
+			_, _, err = installer.ScaleUp(masterIpList, nodeIpList)
 
 			if err = cf.SaveAll(); err != nil {
 				return err
@@ -136,8 +136,8 @@ func NewJoinCmd() *cobra.Command {
 	joinCmd.Flags().StringVar(&joinArgs.Pk, "pk", common.GetHomeDir()+"/.ssh/id_rsa", "set baremetal server private key")
 	joinCmd.Flags().StringVar(&joinArgs.PkPassword, "pk-passwd", "", "set baremetal server private key password")
 	joinCmd.Flags().StringSliceVarP(&joinArgs.CustomEnv, "env", "e", []string{}, "set custom environment variables")
-	joinCmd.Flags().IPSliceVarP(&newMasters, "masters", "m", nil, "set Count or IPList to masters")
-	joinCmd.Flags().IPSliceVarP(&newWorkers, "nodes", "n", nil, "set Count or IPList to nodes")
+	joinCmd.Flags().StringVarP(&newMasters, "masters", "m", "", "set Count or IPList to masters")
+	joinCmd.Flags().StringVarP(&newWorkers, "nodes", "n", "", "set Count or IPList to nodes")
 	joinCmd.Flags().StringVarP(&clusterName, "cluster-name", "c", "", "specify the name of cluster")
 	return joinCmd
 }
